refactor: Delete `src/gwmock/signal/base.py`
