+++ tests/cli/utils/test_cli_utils_config.py
@@ -206,12 +206,19 @@ class TestConfig:
         config = Config(
             simulators={
                 "noise": SimulatorConfig(class_="WhiteNoise"),
-                "signal": SimulatorConfig(class_="SignalSimulator"),
+                "signal": SimulatorConfig(class_="CBCSignalSimulator"),
             }
         )
         assert len(config.simulators) == NUM_SIMULATORS
         assert "noise" in config.simulators
         assert "signal" in config.simulators
+
+    def test_config_rejects_removed_signal_simulator(self):
+        """Legacy signal simulator configs should direct users to orchestration."""
+        with pytest.raises(
+            ValidationError, match=r"orchestration\.population.*orchestration\.signal.*orchestration\.noise"
+        ):
+            Config(simulators={"signal": SimulatorConfig(class_="SignalSimulator")})
 
     def test_config_serialization(self):
         """Test Config serialization for YAML export."""
@@ -426,9 +433,11 @@ class TestResolveClassPath:
         assert path == "gwmock.noise.WhiteNoise"
 
     def test_resolve_class_path_simple_signal(self):
-        """Test resolving simple class name in signal section."""
-        path = resolve_class_path("SignalSimulator", "signal")
-        assert path == "gwmock.signal.SignalSimulator"
+        """Removed legacy signal class should raise a migration error."""
+        with pytest.raises(
+            ValueError, match=r"Legacy 'simulators\.signal\.class: SignalSimulator' is no longer supported"
+        ):
+            resolve_class_path("SignalSimulator", "signal")
 
     def test_resolve_class_path_full_path(self):
         """Test resolving full import path."""
@@ -549,6 +558,12 @@ class TestValidateConfig:
         """Test validation fails when output field is invalid."""
         config = {"simulators": {"noise": {"class": "WhiteNoise", "output": "invalid"}}}
         with pytest.raises(ValueError, match="'output' must be a dictionary"):
+            validate_config(config)
+
+    def test_validate_config_rejects_removed_signal_simulator(self):
+        """Legacy signal simulator configs should raise a migration error."""
+        config = {"simulators": {"signal": {"class": "SignalSimulator"}}}
+        with pytest.raises(ValueError, match=r"orchestration\.population.*orchestration\.signal.*orchestration\.noise"):
             validate_config(config)
 
     def test_validate_config_invalid_globals_field(self):
@@ -781,7 +796,7 @@ simulators:
                     arguments={"seed": 42},
                 ),
                 "signal": SimulatorConfig(
-                    class_="SignalSimulator",
+                    class_="CBCSignalSimulator",
                     arguments={"mass1": 10, "mass2": 10},
                 ),
             },
